backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
dj-static==0.0.6
Django==3.1.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
static3==0.7.0
